Issue #26: Work around precision errors in feasible.point()
This fix prevents lpSolve() hanging on certain input in which the
numbers are of very different orders of magnitude. This is known to
cause problems with many (most?) solvers.  In the example, one element
of const.mat was 1e-12, and an element of const.rhs was -800,000. By
replacing const.mat with a version rounded to 6 decimal places, the
problem is avoided. Thanks to Sam Buttrey, lp maintainer, for this
advice.