Issue #30: Move qhull docs to vignettes
The devel version of R (as of 2019/02/11) wants to delete the inst/doc
directory when building, in preparation for putting vignettes
there. This breaks the Qhull docs, which are in inst/doc. To work
around this, the qhull docs need to be moved to vignettes/qhull, and
included using vignettes/.install_extras. Various links need to be
changed too. See r-lib/pkgbuild#58 for explanation of the behaviour.