Issue #29: Make delaunayn() and convhulln() consistent
There was a problem with replacing full by suggesting using "FA Fn"
for options in delaunayn, since this got rid of the default options,
which could lead to unrpedictable results. The new solution is to make
a new output.options argument to delaunayn() and convhulln().  This
takes Qhull options, but can also be set to TRUE to specify all
possible output options.