Issue #44:  Open R_tmpnam()/free() causes crash in TERR
Bill Dunlap (https://github.com/BillDunlap) has supplied this patch,
with the following explanation:

Because TERR is not compiled with gcc (it uses either the Intel or
Microsoft C++ compiler) on Windows, calling free() in a package's DLL
for memory allocated with malloc() in TERR's own DLL causes a
crash. The R API function R_tmpnam() causes this problem in the
geometry package. To avoid it one could compute tempfile() twice in
the R code, computing filenames tmp_stdout and tmp_stderr, instead of
passing tempdir() to the C code and calling R_tmpnam() twice (and
free() twice) there. This might also help those adventurous types who
want to compile R with alternate compilers. I've attached a file of
diffs that accomplishes this. All the package's tests pass before and
after the patch.