Issue #12: QH6227 impossible to triangulate a dataset
Following the upgrade to Qhull 8.0.2, the issue reported in #12
appears to have disappeard on visual inspection. However, there are
some points not included in the triangulation with the default
delaunayn options.

This patch adds code to delaunayn to throw a warning about points
missing from the triangulation, and suggests options that may help, or
else recentring the input data.